Create workers in parallel using ThreadPool.
