+++ deployments/swarm/buildSrc/src/main/java/se/nbis/lega/deployment/cluster/ListTask.java
@@ -4,7 +4,7 @@ import org.gradle.api.tasks.TaskAction;
 
 import java.io.IOException;
 
-public class LSTask extends ClusterTask {
+public class ListTask extends ClusterTask {
 
     @TaskAction
     public void run() throws IOException {
