+++ deployments/swarm/cluster/build.gradle
@@ -1,12 +1,12 @@
 import se.nbis.lega.deployment.cluster.CreateManagerTask
-import se.nbis.lega.deployment.cluster.CreateWorkerTask
+import se.nbis.lega.deployment.cluster.CreateWorkersTask
 import se.nbis.lega.deployment.cluster.EnvTask
-import se.nbis.lega.deployment.cluster.LSTask
+import se.nbis.lega.deployment.cluster.ListTask
 
 task createManager(type: CreateManagerTask)
 
-task createWorker(type: CreateWorkerTask)
+task createWorkers(type: CreateWorkersTask)
 
-task ls(type: LSTask)
+task list(type: ListTask)
 
 task env(type: EnvTask)
